+++ lib/properties_intf.ml
@@ -3013,7 +3013,10 @@ type mask =
   | Revert_layer
   | Var of mask var
 
-type border_image_slice_item = Number of float | Pct of float
+(** CSS Backgrounds 3 sec. 5.2 to 5.4 write the numeric halves of the
+    border-image slots as [<number [0,inf]>], which a [calc()] satisfies, so
+    each carries a {!type-number} rather than a float. *)
+type border_image_slice_item = Number of number | Pct of float
 
 type border_image_slice_offsets = {
   offsets : border_image_slice_item list;
@@ -3032,12 +3035,12 @@ type border_image_slice =
   | Var of border_image_slice var
 
 type border_image_width_item =
-  | Number of float
+  | Number of number
   | Pct of float
   | Length of length
   | Auto
 
-type border_image_outset_item = Number of float | Length of length
+type border_image_outset_item = Number of number | Length of length
 type border_image_repeat_keyword = Stretch | Repeat | Round | Space
 
 (* CSS Backgrounds 3 sec. 5.5: [border-image-repeat] is one or two keywords